requirements.txt
# Core Web Framework
Flask==2.2.2
gunicorn==20.1.0

# Flask Dependencies
click==8.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2

# Discord Bot
discord.py==2.4.0

# Google API & Authentication
google-auth>=2.15.0,<3.0.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.34.0

# HTTP Requests and Parsing
requests==2.26.0
beautifulsoup4==4.12.3

# Timezone Handling
pytz==2023.3
google-generativeai==0.4.0

# System info
psutil==5.9.5

# Image generation
Pillow==11.1.0

# Reddit Fetching
asyncpraw==7.8.1
aiohttp==3.11.6
